requirements.txt
discord.py>=2.3.0
google-cloud-compute>=1.14.0
python-dotenv>=1.0.0
